remove: 기존 RDB 기반 리프레시 토큰 로직 삭제 #60
