--- src/main/java/com/parkez/auth/authentication/refresh/RefreshTokenRepository.java
@@ -1,12 +1,0 @@
-package com.parkez.auth.authentication.refresh;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-
-	void deleteByUserId(Long userId);
-
-	boolean existsByUserId(Long userId);
-
-	boolean existsByToken(String refreshToken);
-}
