+++ src/main/java/com/parkez/auth/service/TokenManager.java
@@ -3,7 +3,7 @@ package com.parkez.auth.service;
 import org.springframework.stereotype.Service;
 
 import com.parkez.auth.authentication.jwt.JwtProvider;
-import com.parkez.auth.authentication.refresh.RefreshTokenRedisStore;
+import com.parkez.auth.authentication.refresh.RefreshTokenStore;
 import com.parkez.auth.dto.response.TokenResponse;
 import com.parkez.auth.exception.AuthErrorCode;
 import com.parkez.common.exception.ParkingEasyException;
@@ -16,7 +16,7 @@ import lombok.RequiredArgsConstructor;
 public class TokenManager {
 
 	private final JwtProvider jwtProvider;
-	private final RefreshTokenRedisStore refreshTokenRedisStore;
+	private final RefreshTokenStore refreshTokenStore;
 
 	public TokenResponse issueTokens(User user) {
 		String accessToken = jwtProvider.createAccessToken(user.getId(), user.getEmail(),
@@ -24,7 +24,7 @@ public class TokenManager {
 
 		String refreshToken = jwtProvider.createRefreshToken(user.getId());
 
-		refreshTokenRedisStore.set(user.getId(),refreshToken);
+		refreshTokenStore.set(user.getId(),refreshToken);
 
 		return TokenResponse.of(accessToken, refreshToken);
 	}
@@ -34,7 +34,7 @@ public class TokenManager {
 	}
 
 	public void validateRefreshToken(Long userId) {
-		if (!refreshTokenRedisStore.existsBy(userId)) {
+		if (!refreshTokenStore.existsBy(userId)) {
 			throw new ParkingEasyException(AuthErrorCode.TOKEN_NOT_FOUND);
 		}
 	}
